Fix for JuliaLang/julia@fc05d74d0bd2117c8691e68e751ee9b212a5d3a2

+++ src/NIfTI.jl
@@ -90,18 +90,18 @@ end
 const SIZEOF_HDR = int32(348)
 
 const NIfTI_DT_BITSTYPES = (Int16=>Type)[
-	2 => Uint8,
-	4 => Int16,
-	8 => Int32,
-	16 => Float32,
-	32 => Complex64,
-	64 => Float64,
-	256 => Int8,
-	512 => Uint16, 
-	768 => Uint32,
-	1024 => Int64,
-	1280 => Uint64,
-	1792 => Complex128
+	int16(2) => Uint8,
+	int16(4) => Int16,
+	int16(8) => Int32,
+	int16(16) => Float32,
+	int16(32) => Complex64,
+	int16(64) => Float64,
+	int16(256) => Int8,
+	int16(512) => Uint16, 
+	int16(768) => Uint32,
+	int16(1024) => Int64,
+	int16(1280) => Uint64,
+	int16(1792) => Complex128
 ]
 const NIfTI_DT_BITSTYPES_REVERSE = (Type=>Int16)[]
 for (k, v) in NIfTI_DT_BITSTYPES
